Development (#38)
* init

* - added env file

* - removed loans menu and transfer it under Transactions
- added AddUpdateUserDrawer

* - bug fixes
- added role CRUD
- added permissions CRUD

* added roles and bug fixes

* - added user management
- bug fixes

* - fixed display of branches when editting a user

* - added groups management

* - fixed delete in all modules
- added client management
- bug fixes

* - added loans

* - bug fixes
- added roles for nav and data related
- added details based on the last meeting

* - bug fixes and enhancements

* - added loan applications
- bug fixes
- permissions
and many more...

* - added client detail page
- added modal
- bug fixes

* bug fixes

* updates

* updates

* bug fixes

* - updated the login screen

* - bug fixes

* Stopped tracking .env File

* Stopped tracking next.config.js File

* - added profile picture for clients

* - bug fixes

* - added daily cash collections
- bug fixes

* - fixed error in client list

* - bug fixes

* - bug fixes

* - bug fixes

* bug fixes

* - updates

* fixed issues

* - updates

* - updates

* updates

* updates

* bug fixed

* - updates

* - fixed issue

* - updates

* - updates

* - bug fix

* - bug fixes

* - bug fixes

* - bug fix

* - updates and bug fixes

* - bug fix

* - bug fixes

* - remarks update

* - updates

* - updates

* - fixes

* - debugging

* - removing debug

* - bug fix

* - bug fixed

* - update query by branch/group

* - page size fixes

* - updates

* - updates
- bug fixes

* - fixes

+++ src/pages/api/clients/delete.js
@@ -13,22 +13,32 @@ async function deleteUser(req, res) {
   let statusCode = 200;
   let response = {};
 
-  const clients = await db.collection("client").find({ _id: ObjectId(_id) }).toArray();
+  const loan = await db.collection('loans').find({ clientId: _id }).toArray();
 
-  if (clients.length > 0) {
-    await db
-        .collection('client')
-        .deleteOne({ _id: ObjectId(_id) });
-
-    response = {
-      success: true,
-    };
-  } else {
+  if (loan.length > 0) {
     response = {
       error: true,
       fields: ["id"],
-      message: `Client with the id "${_id}" not exists`,
+      message: `Error. Client has a loan already.`,
     };
+  } else {
+    const clients = await db.collection("client").find({ _id: ObjectId(_id) }).toArray();
+
+    if (clients.length > 0) {
+      await db
+          .collection('client')
+          .deleteOne({ _id: ObjectId(_id) });
+
+      response = {
+        success: true,
+      };
+    } else {
+      response = {
+        error: true,
+        fields: ["id"],
+        message: `Client with the id "${_id}" not exists`,
+      };
+    }
   }
 
   res
